Update install virtual test
