+++ packages/node/test/integration/install-virtual.spec.ts
@@ -1,27 +1,17 @@
 import { Node as NodeTypes } from "@counterfactual/types";
-import { Wallet } from "ethers";
 import { One, Zero } from "ethers/constants";
-import {
-  JsonRpcProvider,
-  Provider,
-  TransactionRequest
-} from "ethers/providers";
-import { parseEther } from "ethers/utils";
-import { fromMnemonic } from "ethers/utils/hdnode";
-import { v4 as generateUUID } from "uuid";
 
-import { IMessagingService, IStoreService, Node, NodeConfig } from "../../src";
+import { Node } from "../../src";
 import { APP_INSTANCE_STATUS } from "../../src/db-schema";
-import { MNEMONIC_PATH } from "../../src/signer";
 import {
   CreateChannelMessage,
   InstallVirtualMessage,
   NODE_EVENTS,
   ProposeVirtualMessage
 } from "../../src/types";
-import { CF_PATH } from "../global-setup.jest";
 import { LocalFirebaseServiceFactory } from "../services/firebase-server";
 
+import { setup } from "./setup";
 import {
   collateralizeChannel,
   confirmProposedVirtualAppInstanceOnNode,
@@ -37,72 +27,16 @@ describe("Node method follows spec - proposeInstallVirtual", () => {
   jest.setTimeout(35000);
 
   let firebaseServiceFactory: LocalFirebaseServiceFactory;
-  let messagingService: IMessagingService;
   let nodeA: Node;
-  let storeServiceA: IStoreService;
   let nodeB: Node;
-  let storeServiceB: IStoreService;
   let nodeC: Node;
-  let storeServiceC: IStoreService;
-  let nodeConfig: NodeConfig;
-  let provider: JsonRpcProvider;
 
   beforeAll(async () => {
-    firebaseServiceFactory = new LocalFirebaseServiceFactory(
-      process.env.FIREBASE_DEV_SERVER_HOST!,
-      process.env.FIREBASE_DEV_SERVER_PORT!
-    );
-    messagingService = firebaseServiceFactory.createMessagingService(
-      process.env.FIREBASE_MESSAGING_SERVER_KEY!
-    );
-    nodeConfig = {
-      STORE_KEY_PREFIX: process.env.FIREBASE_STORE_PREFIX_KEY!
-    };
-
-    provider = new JsonRpcProvider(global["ganacheURL"]);
-
-    storeServiceA = firebaseServiceFactory.createStoreService(
-      process.env.FIREBASE_STORE_SERVER_KEY! + generateUUID()
-    );
-
-    // generate new mnemonics so owner addresses are different for creating
-    // a channel in this suite
-    const { A_MNEMONIC, B_MNEMONIC } = await generateNewFundedMnemonics(
-      global["fundedPrivateKey"],
-      provider
-    );
-
-    storeServiceA.set([{ key: MNEMONIC_PATH, value: A_MNEMONIC }]);
-    nodeA = await Node.create(
-      messagingService,
-      storeServiceA,
-      nodeConfig,
-      provider,
-      global["networkContext"]
-    );
-
-    storeServiceB = firebaseServiceFactory.createStoreService(
-      process.env.FIREBASE_STORE_SERVER_KEY! + generateUUID()
-    );
-    storeServiceB.set([{ key: MNEMONIC_PATH, value: B_MNEMONIC }]);
-    nodeB = await Node.create(
-      messagingService,
-      storeServiceB,
-      nodeConfig,
-      provider,
-      global["networkContext"]
-    );
-
-    storeServiceC = firebaseServiceFactory.createStoreService(
-      process.env.FIREBASE_STORE_SERVER_KEY! + generateUUID()
-    );
-    nodeC = await Node.create(
-      messagingService,
-      storeServiceC,
-      nodeConfig,
-      provider,
-      global["networkContext"]
-    );
+    const result = await setup(global, true, true);
+    nodeA = result.nodeA;
+    nodeB = result.nodeB;
+    nodeC = result.nodeC!;
+    firebaseServiceFactory = result.firebaseServiceFactory;
   });
 
   afterAll(() => {
@@ -253,35 +187,3 @@ describe("Node method follows spec - proposeInstallVirtual", () => {
     }
   );
 });
-
-async function generateNewFundedMnemonics(
-  fundedPrivateKey: string,
-  provider: Provider
-) {
-  const fundedWallet = new Wallet(fundedPrivateKey, provider);
-  const A_MNEMONIC = Wallet.createRandom().mnemonic;
-  const B_MNEMONIC = Wallet.createRandom().mnemonic;
-
-  const signerAPrivateKey = fromMnemonic(A_MNEMONIC).derivePath(CF_PATH)
-    .privateKey;
-  const signerBPrivateKey = fromMnemonic(B_MNEMONIC).derivePath(CF_PATH)
-    .privateKey;
-
-  const signerAAddress = new Wallet(signerAPrivateKey).address;
-  const signerBAddress = new Wallet(signerBPrivateKey).address;
-
-  const transactionToA: TransactionRequest = {
-    to: signerAAddress,
-    value: parseEther("0.1").toHexString()
-  };
-  const transactionToB: TransactionRequest = {
-    to: signerBAddress,
-    value: parseEther("0.1").toHexString()
-  };
-  await fundedWallet.sendTransaction(transactionToA);
-  await fundedWallet.sendTransaction(transactionToB);
-  return {
-    A_MNEMONIC,
-    B_MNEMONIC
-  };
-}
