+++ packages/node/test/integration/setup.ts
@@ -1,11 +1,23 @@
-import { JsonRpcProvider } from "ethers/providers";
+import { Wallet } from "ethers";
+import {
+  JsonRpcProvider,
+  Provider,
+  TransactionRequest
+} from "ethers/providers";
+import { parseEther } from "ethers/utils";
+import { fromMnemonic } from "ethers/utils/hdnode";
 import { v4 as generateUUID } from "uuid";
 
 import { MNEMONIC_PATH, Node } from "../../src";
+import { CF_PATH } from "../global-setup.jest";
 import { LocalFirebaseServiceFactory } from "../services/firebase-server";
 import { A_MNEMONIC, B_MNEMONIC } from "../test-constants.jest";
 
-export async function setup(global: any, nodeCPresent: boolean = false) {
+export async function setup(
+  global: any,
+  nodeCPresent: boolean = false,
+  newMnemonics: boolean = false
+) {
   const firebaseServiceFactory = new LocalFirebaseServiceFactory(
     process.env.FIREBASE_DEV_SERVER_HOST!,
     process.env.FIREBASE_DEV_SERVER_PORT!
@@ -19,11 +31,24 @@ export async function setup(global: any, nodeCPresent: boolean = false) {
 
   const provider = new JsonRpcProvider(global["ganacheURL"]);
 
+  let mnemonicA = A_MNEMONIC;
+  let mnemonicB = B_MNEMONIC;
+  if (newMnemonics) {
+    // generate new mnemonics so owner addresses are different for creating
+    // a channel in this suite
+    const mnemonics = await generateNewFundedMnemonics(
+      global["fundedPrivateKey"],
+      provider
+    );
+    mnemonicA = mnemonics.A_MNEMONIC;
+    mnemonicB = mnemonics.B_MNEMONIC;
+  }
+
   const storeServiceA = firebaseServiceFactory.createStoreService(
     process.env.FIREBASE_STORE_SERVER_KEY! + generateUUID()
   );
 
-  storeServiceA.set([{ key: MNEMONIC_PATH, value: A_MNEMONIC }]);
+  storeServiceA.set([{ key: MNEMONIC_PATH, value: mnemonicA }]);
   const nodeA = await Node.create(
     messagingService,
     storeServiceA,
@@ -35,7 +60,7 @@ export async function setup(global: any, nodeCPresent: boolean = false) {
   const storeServiceB = firebaseServiceFactory.createStoreService(
     process.env.FIREBASE_STORE_SERVER_KEY! + generateUUID()
   );
-  storeServiceB.set([{ key: MNEMONIC_PATH, value: B_MNEMONIC }]);
+  storeServiceB.set([{ key: MNEMONIC_PATH, value: mnemonicB }]);
   const nodeB = await Node.create(
     messagingService,
     storeServiceB,
@@ -61,4 +86,36 @@ export async function setup(global: any, nodeCPresent: boolean = false) {
   }
 
   return { nodeA, nodeB, firebaseServiceFactory };
+}
+
+async function generateNewFundedMnemonics(
+  fundedPrivateKey: string,
+  provider: Provider
+) {
+  const fundedWallet = new Wallet(fundedPrivateKey, provider);
+  const A_MNEMONIC = Wallet.createRandom().mnemonic;
+  const B_MNEMONIC = Wallet.createRandom().mnemonic;
+
+  const signerAPrivateKey = fromMnemonic(A_MNEMONIC).derivePath(CF_PATH)
+    .privateKey;
+  const signerBPrivateKey = fromMnemonic(B_MNEMONIC).derivePath(CF_PATH)
+    .privateKey;
+
+  const signerAAddress = new Wallet(signerAPrivateKey).address;
+  const signerBAddress = new Wallet(signerBPrivateKey).address;
+
+  const transactionToA: TransactionRequest = {
+    to: signerAAddress,
+    value: parseEther("0.1").toHexString()
+  };
+  const transactionToB: TransactionRequest = {
+    to: signerBAddress,
+    value: parseEther("0.1").toHexString()
+  };
+  await fundedWallet.sendTransaction(transactionToA);
+  await fundedWallet.sendTransaction(transactionToB);
+  return {
+    A_MNEMONIC,
+    B_MNEMONIC
+  };
 }
